+++ src/Products/PloneMeeting/tests/testPerformances.py
@@ -378,7 +378,7 @@ class testPerformances(PloneMeetingTestCase):
         for time in range(times):
             get_organizations(not_empty_suffix='advisers', caching=caching)
 
-    def test_pm_GetMeetingConfig(self):
+    def test_pm_SpeedGetMeetingConfig(self):
         '''Test ToolPloneMeeting.getMeetingConfig method performances.
            We call the method 2000 times, this is what happens when displaying
            a meeting containing 100 items.'''
@@ -686,6 +686,27 @@ class testPerformances(PloneMeetingTestCase):
     def _duplicateItem(self, item, copyAnnexes=False):
         '''Helper method that actually duplicated given p_item.'''
         item.clone(copyAnnexes=copyAnnexes)
+
+    def test_pm_SpeedGetMeeting(self):
+        '''Test MeetingItem.getMeeting method performances.
+           We call the method 2000 times, this is what happens when displaying
+           a dashboard of 100 items.'''
+        self.changeUser('pmManager')
+        self.create('Meeting', date=DateTime('2020/11/23'))
+        item = self.create('MeetingItem')
+        self.presentItem(item)
+        # call getMeeting 2000 times wihout caching
+        self._getMeetingOnItem(item, 2000, caching=False)
+        # call getMeeting 2000 times with caching
+        self._getMeetingOnItem(item, 2000, caching=True)
+
+    @timecall
+    def _getMeetingOnItem(self, item, times=1, caching=True):
+        ''' '''
+        pm_logger.info(
+            'Call {0} times, with caching={1}'.format(times, caching))
+        for time in range(times):
+            item.getMeeting(only_uid=False, caching=caching)
 
 
 def test_suite():